requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
mlflow>=2.7.0
joblib>=1.3.0
python-dotenv>=1.0.0